[eviction] Move SelectNodeToEvict into EvictionManager
