+++ src/test/net_peer_eviction_tests.cpp
@@ -4,6 +4,7 @@
 
 #include <netaddress.h>
 #include <net.h>
+#include <node/eviction_impl.h>
 #include <test/util/net.h>
 #include <test/util/setup_common.h>
