+++ src/test/util/net.cpp
@@ -5,7 +5,7 @@
 #include <test/util/net.h>
 
 #include <chainparams.h>
-#include <node/eviction.h>
+#include <node/eviction_impl.h>
 #include <net.h>
 #include <net_processing.h>
 #include <netmessagemaker.h>
